fix: Rename helpers file to makeInputAdornment
Another `helpers` file already exists, unlike the `.js` extension.
As a result, after the build, the contents of the `.jsx` file
were missing, so the `makeInputAdornment` function
no longer existed and caused an error during paper creation.
